Read configuration files into a single dictionary
This change adds support for reading multiple configuration files in
YAML format into a single dictionary. The configuration files can be
specified using command line options (`-c` or `--config` option).

When no configuration file option is specified in the command line, the
framework looks for `config.base.yaml` and `config.yaml` in the current
directory. When multiple configuration files are specified, the
configuration in the file that is specified later is merged with the
configuration in the files specified earlier.