Measure code coverage in subprocesses
Prior to this change, the following commands in `Makefile` do not
measure code coverage in subprocesses.

    coverage run --source . --branch -m unittest discover -v
    coverage report --show-missing
    coverage html

The coverage results can be seen on the terminal as the output of
`coverage report --show-missing` and as HTML in `htmlcov/index.html`.

With the above commands, we see missed coverage for lines
of code in `mockcloud.read`, `mockcheck.eval`, `filestore.write`, and
`filestore.done`, even though these code lines were exercised indirectly
from `cloudmarker/test/test_main.py`.

This issue occurs because these lines of code are not executed within
the main process. Instead, they are exercised from the workers launched
as separate processes using `multiprocessing.Process`. The `coverage`
tool does not measure code execution in subprocesses by default. It has
to be run with `--concurrency=multiprocessing` to measure code execution
in subprocesses. But doing so leads to another issue:

    $ coverage run --source=. --concurrency=multiprocessing --branch \
      -m unittest discover -v
    Options affecting multiprocessing must be specified in a
    configuration file.
    Use 'coverage help' for help.

If we use `--concurrency` in the command line option, other options like
`--source` and `--branch` must be specified in a configuration file
(`.coveragerc` by default). This is the reason why `.coveragerc` has
been introduced in this change.

See comments by Ned Batchelder, the author of `coverage`, at
<https://stackoverflow.com/a/52379777/303363> for some more information
about this.

This simplifies the `coverage` command in the `Makefile` because most of
the command line options have now moved to `.coveragerc`.

Additionally, in favour of simplifying the `coverage` command further,
the `discover` argument to `unittest` module has been removed in the
`Makefile` because test discovery is the default behaviour of the
`unittest` module. Here is the relevant excerpt from
<https://docs.python.org/3/library/unittest.html>:

> When executed without arguments Test Discovery is started:
>
>     python -m unittest