Fix invalid JSON file written by FileStore plugin
Prior to this fix, the FileStore plugin writes an invalid JSON file to
the file system. Here is an example file written by this plugin prior
to this fix:

	[
	{
	  "record_num": 0,
	  "record_type": "foo"
	},
	{
	  "record_num": 2,
	  "record_type": "foo"
	},
	{
	  "record_num": 4,
	  "record_type": "foo"
	},
	{
	  "record_num": 6,
	  "record_type": "foo"
	},
	{
	  "record_num": 8,
	  "record_type": "foo"
	},
	]

The error here is the trailing comma at the end of all items in the JSON
array. This issue occurs because the plugin adds a comma after every
item. This change fixes this issue.