Insert records into MongoDB in arbitrary order
This change overrides the default parameter of `ordered=True` to
`ordered=False` so that we do not insert records into MongoDB serially,
and instead we insert them in arbitrary order. This ensures that if a
record insertion fails, the other valid records are still inserted in
parallel.

Additionally, this change introduces error handling for exceptions
caused while insertion of records.