Add mongodbstore key in config.base.yaml
Provide a `mongodbstore` key in `config.base.yaml` that would create a
MongoDBStore plugin object with its default parameters. This would allow
users and developers to use this key readily in the user configuration
(e.g., `config.yaml`) as long as they intend to connect to a locally
running MongoDB that does not require authentication.