Write help for -c/--config in imperative mood
There is a convention of writing help messages using `argparse` in
imperative mood. This convention is implied by the default help message
for `-h`/`--help` option defined in `argparse`.

    $ python3 -m cloudmarker -h
    usage: cloudmarker [-h] [-c CONFIG [CONFIG ...]] [-n]

    optional arguments:
      -h, --help            show this help message and exit
      ...

Therefore, write the help messages for all the options in the same way.
Currently, only the help message for `-c`/`--config` does not conform to
this convention, so this commit fixes that.