Remove EmailAlert configuration from base config
The `baseconfig.py` module contains configuration for plugins that can
be used readily in a local development environment or in an environment
where the store or alert destinations are present locally. For these
plugins to be used readily, they should have sensible defaults.

Not all parameters of `EmailAlert` plugin can be defined with sensible
defaults. For example, there are no sensible defaults for `from_addr`
and `to_addrs`. Therefore, there should be no configuration for this in
`baseconfig.py`.