Make EmailAlert docs and code consistent
The EmailAlert docstrings have been updated to make it look good from a
user's perspective. In some places, the docstrings have been fixed to
describe the functionality in the code correctly. In some other cases,
the code has been updated to be consistent with the updated docstrings.

The following changes have been made to the plugin parameters:

  - Rename `sender` to `from_addr`.
  - Rename `to` to `to_addrs`.
  - Remove `body`.
  - Add defaults for `subject`, `host`, and `port`.

The names `from_addr` and `to_addrs` have been picked from the Python
`smtplib` documentation. They have done the work of choosing good names,
so we just reuse it. The `to_addrs` name has an additional advantage
that it hints at the value being a list of addresses.

The `body` parameter has been removed because this is not a generic
emailer class. This is an alert plugin which will decide its own email
content based on the events it receives in its `write()` method.
There is never a situation where the email content can be specified via
configuration.

The defaults for `host` and `port` have also been picked from the Python
`smtplib` documentation. They implement sensible default behaviour for
these defaults, so we reuse that behaviour by choosing the same
defaults.