Do not alter cloud record in FirewallRuleEvent
Prior to this commit, `FirewallRuleEvent` plugin has a bug that caused
it to change the extended `record_type` property of the cloud record to
`firewall_rule_event`. This happened because this plugin first copied
the reference to cloud record's `ext` bucket to the new event record and
then set its `record_type` property in the `ext` bucket.

This fix performs a deepcopy of the `ext` bucket from the cloud record
to the event record to ensure that altering the `ext` bucket of the
event record does not alter that of the cloud record.