Do not disable cache test dist installation
Ignoring the pip cache led to longer test installation time. Therefore,
stop ignoring it to improve the installation time.

The pip cache was ignored earlier because sometimes pip picked up an
older version of a package for this project. Recently, we found that it
happens sometimes (such as soon after an upload) despite disabling
cache. Therefore, disabling cache does not offer significant benefits
but increases the test install time.