Update indentation in API docs source files
The indentation style in API documentation source files generated by the
latest version of `sphinx-apidoc` has changed. With the earlier Sphinx
2.0.1, the automodule directives were indented as follows:

	.. automodule:: cloudmarker.alerts
		:members:
		:undoc-members:
		:show-inheritance:

Since Sphinx 2.1.0 and later, they are now indented as follows:

	.. automodule:: cloudmarker.alerts
	   :members:
	   :undoc-members:
	   :show-inheritance:

This change updates the Sphinx project's source code files with this
updated indentation style, so that subsequent running of `sphinx-apidoc`
does not lead to diffs due to indentation changes.