Add AzLogProfileRetentionEvent plugin
The `AzLogProfileRetentionEvent` plugin evaluate an Azure `log_profile`
record and generate and event of type `log_profile_retention_event` if
the log profile's retention policy is configured for less number of days
than required. The default value for minumum number of days is 365, as
per CIS recommendation but can be changed to any other desired value
through the `min_retention_days` parameter.

It is to be noted that setting retention (days) to 0 from portal
sets days to 0 and enabled to false, which will retain the data forever.