Add AzLogProfileMissingLocationEvent plugin
An Azure log profile must capture events for all the available
locations, so that activity logs from all the locations are available
for incident response and investigation.

As many of the events in the activity logs are global events, a log
profile must also be enabled for `global` location.

This event plugin evaluates records of type `log_profile` and generates
an event of type `log_profile_missing_location_event` if a log profile
is not enabled for any of the available location/region for that
subscription including `global`.