Upgrade google-auth from 1.11.2 to 1.17.0
We need to upgrade rsa from 4.0 to 4.7 resolve a few vulnerabilities.
However, the current dependency of google-auth==1.11.2 prevents
rsa-4.7 from being installed. For example, here is the error we get in
Python 3.8

    ERROR: Cannot install -r requirements.txt (line 21) and rsa==4.7
    because these package versions have conflicting dependencies.

    The conflict is caused by:
        The user requested rsa==4.7
        google-auth 1.11.2 depends on rsa<4.1 and >=3.1.4

Version 1.11.2 of google-auth indeed did not support rsa version 4.1
or later. Version 1.17.0 is the earliest version of google-auth to
support rsa version 4.1 or later. See the following URLs to confirm
this:

  - https://github.com/googleapis/google-auth-library-python/blob/v1.11.2/setup.py#L24
  - https://github.com/googleapis/google-auth-library-python/blob/v1.17.0/setup.py#L24

This change upgrades the google-auth library to 1.17.0, so that
rsa-4.7 can be installed without a conflict.

Lines starting # with '#' will be ignored, and an empty message aborts
the commit. # # On branch dependabot/pip/rsa-4.7 # Your branch is up
to date with 'origin/dependabot/pip/rsa-4.7'. # # Changes to be
committed: # modified: requirements.txt #